import ops._ within macros
- Move functions out to PickleFunctions trait.
- import ops._ from macros since macro implementations use ops.
- Add test to check you can generate pickler without having opts in the
call site scope.
